requirements.txt
python-pptx
pillow
pandas
openpyxl
xlrd
Gooey